done with checking and handling the map

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: welim <welim@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/18 15:38:49 by welim             #+#    #+#             */
-/*   Updated: 2022/04/18 15:55:52 by welim            ###   ########.fr       */
+/*   Updated: 2023/11/23 17:59:54 by welim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
